+++ tx.go
@@ -47,7 +47,7 @@ func ProposePayment(
 		txOpts = &o
 	}
 
-	contractAddr, deployTx, con, err := contract.DeployTredd(txOpts, client, seller, tokenType, amount, collateral, clearRoot, cipherRoot, revealDeadline.Unix(), refundDeadline.Unix())
+	contractAddr, deployTx, con, err := contract.DeployTredd(txOpts, client, seller, tokenType, amount, collateral, clearRoot, cipherRoot, uint64(revealDeadline.Unix()), uint64(refundDeadline.Unix()))
 	if err != nil {
 		return common.Address{}, nil, nil, errors.Wrap(err, "deploying contract")
 	}
@@ -155,8 +155,8 @@ func RevealKey(
 	if err != nil {
 		return nil, nil, errors.Wrap(err, "getting mRevealDeadline")
 	}
-	gotRevealDeadline := time.Unix(gotRevealDeadlineSecs, 0)
-	if gotRevealDeadlineSecs != wantRevealDeadline.Unix() { // lop off fractional seconds from wantRevealDeadline
+	gotRevealDeadline := time.Unix(int64(gotRevealDeadlineSecs), 0)
+	if int64(gotRevealDeadlineSecs) != wantRevealDeadline.Unix() { // lop off fractional seconds from wantRevealDeadline
 		return nil, nil, fmt.Errorf("reveal deadline is %s, want %s", gotRevealDeadline, wantRevealDeadline)
 	}
 
@@ -168,8 +168,8 @@ func RevealKey(
 	if err != nil {
 		return nil, nil, errors.Wrap(err, "getting mRefundDeadline")
 	}
-	if gotRefundDeadlineSecs != wantRefundDeadline.Unix() { // lop off fractional seconds from wantRefundDeadline
-		return nil, nil, fmt.Errorf("refund deadline is %s, want %s", time.Unix(gotRefundDeadlineSecs, 0), wantRefundDeadline)
+	if int64(gotRefundDeadlineSecs) != wantRefundDeadline.Unix() { // lop off fractional seconds from wantRefundDeadline
+		return nil, nil, fmt.Errorf("refund deadline is %s, want %s", time.Unix(int64(gotRefundDeadlineSecs), 0), wantRefundDeadline)
 	}
 
 	paidAmount, err := con.Paid(callOpts)
